+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/iscored/IScoredService.java
@@ -6,11 +6,13 @@ import de.mephisto.vpin.commons.fx.notifications.NotificationFactory;
 import de.mephisto.vpin.connectors.iscored.GameRoom;
 import de.mephisto.vpin.connectors.iscored.IScored;
 import de.mephisto.vpin.connectors.iscored.IScoredGame;
+import de.mephisto.vpin.connectors.iscored.IScoredResult;
 import de.mephisto.vpin.connectors.mania.model.Account;
 import de.mephisto.vpin.connectors.mania.model.TableScore;
 import de.mephisto.vpin.connectors.mania.model.Tournament;
 import de.mephisto.vpin.connectors.mania.model.TournamentTable;
 import de.mephisto.vpin.restclient.PreferenceNames;
+import de.mephisto.vpin.restclient.highscores.logging.SLOG;
 import de.mephisto.vpin.restclient.notifications.NotificationSettings;
 import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.games.GameService;
@@ -54,6 +56,7 @@ public class IScoredService implements PreferenceChangedListener, InitializingBe
         if (gameRoom != null) {
           if (!gameRoom.getSettings().isPublicScoreEnteringEnabled()) {
             LOG.warn("Cancelling iScored score submission, public score submissions are not enabled!");
+            SLOG.warn("Cancelling iScored score submission, public score submissions are not enabled!");
             return;
           }
 
@@ -63,11 +66,13 @@ public class IScoredService implements PreferenceChangedListener, InitializingBe
           IScoredGame gameRoomGame = gameRoom.getGameByVps(vpsTableId, vpsVersionId);
           if (gameRoomGame == null) {
             LOG.info("Skipped iScored score submission, because no game was found for " + tournament);
+            SLOG.info("Skipped iScored score submission, because no game was found for " + tournament);
             return;
           }
 
           if (gameRoomGame.isDisabled()) {
             LOG.info("Skipped iScored score submission, because table " + gameRoomGame + " has disabled flag set.");
+            SLOG.info("Skipped iScored score submission, because table " + gameRoomGame + " has disabled flag set.");
             return;
           }
 
@@ -77,6 +82,7 @@ public class IScoredService implements PreferenceChangedListener, InitializingBe
     }
     catch (Exception e) {
       LOG.error("Failed to load game room from " + tournament.getDashboardUrl() + ": " + e.getMessage(), e);
+      SLOG.error("Failed to load game room from " + tournament.getDashboardUrl() + ": " + e.getMessage());
     }
   }
 
@@ -92,10 +98,12 @@ public class IScoredService implements PreferenceChangedListener, InitializingBe
         String playerName = newScore.getPlayer() != null ? newScore.getPlayer().getName() : newScore.getPlayerInitials();
         if (iScoredGame.isDisabled()) {
           LOG.info("Skipped iScored score submission, because table " + iScoredGame + " has disabled flag set.");
+          SLOG.info("Skipped iScored score submission, because table " + iScoredGame + " has disabled flag set.");
           return;
         }
 
-        IScored.submitScore(gameRoom, iScoredGame, playerName, newScore.getPlayerInitials(), (long) newScore.getNumericScore());
+        IScoredResult result = IScored.submitScore(gameRoom, iScoredGame, playerName, newScore.getPlayerInitials(), (long) newScore.getNumericScore());
+        SLOG.info(result.toString());
         if (Features.NOTIFICATIONS_ENABLED && notificationSettings.isiScoredNotification()) {
           Game game = gameService.getGame(newScore.getGameId());
 
@@ -107,10 +115,12 @@ public class IScoredService implements PreferenceChangedListener, InitializingBe
       }
       else {
         LOG.warn("Failed to get iScored game from game room (" + url + ") for VPS id '" + vpsTableId + "'");
+        SLOG.warn("Failed to get iScored game from game room (" + url + ") for VPS id '" + vpsTableId + "'");
       }
     }
     else {
       LOG.warn("No iScored game room found for " + url);
+      SLOG.warn("No iScored game room found for " + url);
     }
   }